feat: Add basic Node API (#221)
